+++ README-templates.md
@@ -1,0 +1,8 @@
+# File Templates
+
+`workflow-file-templates.zip` can be imported into Android Studio / IntelliJ IDEA to add a few Workflow-specific file templates, via _File > Manage IDE Settings > Import Settings…_.
+
+To update the templates:
+
+* edit them in the IDE (_Settings > Editor > File and Code Templates_)
+* export them (_File > Manage IDE Settings > Import Settings…_), taking care to clear every checkbox except that for File Templates.
