+++ samples/tutorial/tutorial-final/src/main/java/workflow/tutorial/TodoEditScreen.kt
@@ -2,8 +2,12 @@ package workflow.tutorial
 
 import com.squareup.workflow1.ui.AndroidScreen
 import com.squareup.workflow1.ui.ScreenViewFactory
+import com.squareup.workflow1.ui.ScreenViewRunner
 import com.squareup.workflow1.ui.TextController
+import com.squareup.workflow1.ui.ViewEnvironment
 import com.squareup.workflow1.ui.WorkflowUiExperimentalApi
+import com.squareup.workflow1.ui.control
+import com.squareup.workflow1.ui.navigation.setBackHandler
 import workflow.tutorial.views.databinding.TodoEditViewBinding
 
 @OptIn(WorkflowUiExperimentalApi::class)
@@ -18,4 +22,20 @@ data class TodoEditScreen(
 ) : AndroidScreen<TodoEditScreen> {
   override val viewFactory =
     ScreenViewFactory.fromViewBinding(TodoEditViewBinding::inflate, ::TodoEditScreenRunner)
+}
+
+@OptIn(WorkflowUiExperimentalApi::class)
+private class TodoEditScreenRunner(
+  private val binding: TodoEditViewBinding
+) : ScreenViewRunner<TodoEditScreen> {
+
+  override fun showRendering(
+    rendering: TodoEditScreen,
+    environment: ViewEnvironment
+  ) {
+    binding.root.setBackHandler(rendering.onBackPressed)
+    binding.save.setOnClickListener { rendering.onSavePressed() }
+    rendering.title.control(binding.todoTitle)
+    rendering.note.control(binding.todoNote)
+  }
 }
