+++ samples/tutorial/tutorial-final/src/main/java/workflow/tutorial/TodoListScreen.kt
@@ -1,8 +1,13 @@
 package workflow.tutorial
 
+import androidx.recyclerview.widget.LinearLayoutManager
 import com.squareup.workflow1.ui.AndroidScreen
 import com.squareup.workflow1.ui.ScreenViewFactory
+import com.squareup.workflow1.ui.ScreenViewRunner
+import com.squareup.workflow1.ui.ViewEnvironment
 import com.squareup.workflow1.ui.WorkflowUiExperimentalApi
+import com.squareup.workflow1.ui.navigation.setBackHandler
+import workflow.tutorial.views.TodoListAdapter
 import workflow.tutorial.views.databinding.TodoListViewBinding
 
 /**
@@ -21,4 +26,34 @@ data class TodoListScreen(
 ) : AndroidScreen<TodoListScreen> {
   override val viewFactory =
     ScreenViewFactory.fromViewBinding(TodoListViewBinding::inflate, ::TodoListScreenRunner)
+}
+
+@OptIn(WorkflowUiExperimentalApi::class)
+private class TodoListScreenRunner(
+  private val todoListBinding: TodoListViewBinding
+) : ScreenViewRunner<TodoListScreen> {
+
+  private val adapter = TodoListAdapter()
+
+  init {
+    todoListBinding.todoList.layoutManager = LinearLayoutManager(todoListBinding.root.context)
+    todoListBinding.todoList.adapter = adapter
+  }
+
+  override fun showRendering(
+    rendering: TodoListScreen,
+    environment: ViewEnvironment
+  ) {
+    todoListBinding.root.setBackHandler(rendering.onBackPressed)
+    todoListBinding.add.setOnClickListener { rendering.onAddPressed() }
+
+    with(todoListBinding.todoListWelcome) {
+      text =
+        resources.getString(workflow.tutorial.views.R.string.todo_list_welcome, rendering.username)
+    }
+
+    adapter.todoList = rendering.todoTitles
+    adapter.onTodoSelected = rendering.onRowPressed
+    adapter.notifyDataSetChanged()
+  }
 }
